ci: add the buildtest target to run when development
This new target is similar to ci-buildtest, but allows linters to modify files by default. Pass --check to avoid.

We also remove the depedency on CARGO_MAKE_CI, so that we can run the exact same commands and environemnts as CI if we use the ci-buildtest target.

The key difference between ci-buildtest and buildtest is that ci-buildtest enables the -D warnings flag globally while buildtest doesn't. We don't use -D warnings locally to allow local development to build and test with warnings, and to avoid compilation cache invalidation due to change in the compiler flags.